challenge_back/api/serializers.py
from rest_framework import serializers

from import_file.models import Persona


class PersonaSerializer(serializers.ModelSerializer):
    class Meta:
        model = Persona
        fields = '__all__'
