challenge_back/requirements.txt
asgiref==3.3.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.4
django-import-export==2.5.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
greenlet==1.1.0
MarkupPy==1.14
numpy==1.21.0
odfpy==1.4.1
openpyxl==3.0.7
pandas==1.2.5
psycopg2-binary==2.9.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
SQLAlchemy==1.4.19
sqlparse==0.4.1
tablib==3.0.0
xlrd==2.0.1
xlwt==1.3.0
